feat: recognizing commands and updating discord version
