+++ package.json
@@ -4,7 +4,7 @@
     "@discordjs/rest": "^0.3.0",
     "cron": "^1.8.2",
     "discord-api-types": "^0.31.0",
-    "discord.js": "^12.5.3",
+    "discord.js": "^13.6.0",
     "file-system": "^2.2.2",
     "jslint": "^0.12.1",
     "mongoose": "^5.13.14",
